+++ kernel/data-pipeline/core/src/test/java/org/apache/shardingsphere/data/pipeline/api/datanode/JobDataNodeEntryTest.java
@@ -15,9 +15,8 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.test.it.data.pipeline.api.datanode;
+package org.apache.shardingsphere.data.pipeline.api.datanode;
 
-import org.apache.shardingsphere.data.pipeline.api.datanode.JobDataNodeEntry;
 import org.apache.shardingsphere.infra.datanode.DataNode;
 import org.junit.jupiter.api.Test;
