+++ kernel/data-pipeline/core/src/test/resources/META-INF/services/org.apache.shardingsphere.data.pipeline.spi.sqlbuilder.PipelineSQLBuilder
@@ -15,4 +15,5 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.data.pipeline.core.check.consistency.algorithm.fixture.FixturePipelineSQLBuilder
+org.apache.shardingsphere.data.pipeline.core.sqlbuilder.FixturePipelineSQLBuilder
+org.apache.shardingsphere.data.pipeline.core.sqlbuilder.H2PipelineSQLBuilder
